Move CurrentDirectoryProvider to file-system subdirectory
